+++ src/main/java/org/mmarini/fredy2/model/InferenceNode.java
@@ -105,4 +105,6 @@ public interface InferenceNode {
                             .toArray(String[]::new)))),
             List.of("type")
     );
+
+
 }